Add cleanup and restructure some things
