+++ src/Titan.Graphics/D3D12Take2/RenderData.cs
@@ -1,0 +1,9 @@
+using Titan.Core;
+
+namespace Titan.Graphics.D3D12Take2;
+
+public struct RenderData : IResource
+{
+    public ulong FrameCount;
+    public uint FrameIndex;
+}
